+++ src/ephemeris_tools/params.py
@@ -7,6 +7,13 @@ import os
 from dataclasses import dataclass, field
 from typing import TextIO
 
+from ephemeris_tools.constants import (
+    ARCMIN_PER_DEGREE,
+    ARCSEC_PER_DEGREE,
+    DEFAULT_INTERVAL,
+    DEGREES_PER_HOUR_RA,
+)
+
 logger = logging.getLogger(__name__)
 
 # General column IDs (ephem3_xxx.f COL_*)
@@ -307,7 +314,7 @@ def _parse_sexagesimal_to_degrees(text: str, *, is_ra_hours: bool) -> float:
         raise ValueError('Empty angle string')
     try:
         value = float(raw)
-        return value * 15.0 if is_ra_hours else value
+        return value * DEGREES_PER_HOUR_RA if is_ra_hours else value
     except ValueError:
         pass
     parts = [p for p in raw.replace(':', ' ').split() if p]
@@ -317,10 +324,10 @@ def _parse_sexagesimal_to_degrees(text: str, *, is_ra_hours: bool) -> float:
     first = abs(float(parts[0]))
     minutes = float(parts[1]) if len(parts) >= 2 else 0.0
     seconds = float(parts[2]) if len(parts) >= 3 else 0.0
-    value = first + minutes / 60.0 + seconds / 3600.0
+    value = first + minutes / ARCMIN_PER_DEGREE + seconds / ARCSEC_PER_DEGREE
     value *= sign
     if is_ra_hours:
-        value *= 15.0
+        value *= DEGREES_PER_HOUR_RA
     return value
 
 
@@ -432,7 +439,7 @@ def parse_viewer_rings(planet_num: int, tokens: list[str]) -> list[str]:
 
     for token in tokens:
         key = token.strip().lower()
-        if not key:
+        if len(key) == 0:
             continue
         if key == 'none':
             return []
@@ -543,7 +550,7 @@ def parse_column_spec(tokens: list[str]) -> list[int]:
     out: list[int] = []
     for s in tokens:
         s = s.strip()
-        if not s:
+        if len(s) == 0:
             continue
         pref = _int_prefix(s)
         if pref is not None:
@@ -574,7 +581,7 @@ def parse_mooncol_spec(tokens: list[str]) -> list[int]:
     out: list[int] = []
     for s in tokens:
         s = s.strip()
-        if not s:
+        if len(s) == 0:
             continue
         pref = _int_prefix(s)
         if pref is not None:
@@ -638,7 +645,7 @@ def parse_ring_spec(planet_num: int, tokens: list[str]) -> list[int]:
     out: list[int] = []
     for s in tokens:
         s = s.strip()
-        if not s:
+        if len(s) == 0:
             continue
         pref = _int_prefix(s)
         if pref is not None:
@@ -780,7 +787,7 @@ class TrackerParams:
     planet_num: int
     start_time: str
     stop_time: str
-    interval: float = 1.0
+    interval: float = DEFAULT_INTERVAL
     time_unit: str = 'hour'
     observer: Observer = field(default_factory=Observer)
     ephem_version: int = 0
@@ -800,7 +807,7 @@ class EphemerisParams:
     planet_num: int
     start_time: str
     stop_time: str
-    interval: float = 1.0
+    interval: float = DEFAULT_INTERVAL
     time_unit: str = 'hour'
     ephem_version: int = 0
     observer: Observer = field(default_factory=Observer)
@@ -818,15 +825,7 @@ class EphemerisParams:
 
 
 def _safe_float(value: str, default: float) -> float:
-    """Parse string to float; return default on ValueError.
-
-    Parameters:
-        value: String to parse.
-        default: Fallback when parsing fails.
-
-    Returns:
-        Parsed float or default.
-    """
+    """Parse string to float; return default on ValueError."""
     try:
         return float(value)
     except (ValueError, TypeError):
@@ -834,42 +833,27 @@ def _safe_float(value: str, default: float) -> float:
 
 
 def _get_env(key: str, default: str = '') -> str:
-    """Get environment variable, stripped.
-
-    Parameters:
-        key: Environment variable name.
-        default: Value if key missing.
-
-    Returns:
-        Stripped string.
-    """
+    """Get environment variable, stripped."""
     return os.environ.get(key, default).strip()
 
 
 def _get_keys_env(key: str) -> list[str]:
-    """Get repeated env keys (e.g. columns#1, columns#2). Perl/CGI convention.
-
-    Parameters:
-        key: Base name (e.g. 'columns').
-
-    Returns:
-        List of values from key, key#1, key#2, ... or comma/split of single value.
-    """
+    """Get repeated env keys (e.g. columns#1, columns#2). Perl/CGI convention."""
     out: list[str] = []
     i = 1
     while True:
         v = os.environ.get(f'{key}#{i}', '').strip()
-        if not v:
+        if len(v) == 0:
             v = os.environ.get(key if i == 1 else '', '').strip()
-        if not v:
+        if len(v) == 0:
             break
         if '#' in v:
             v = v.split('#')[0].strip()
         out.append(v)
         i += 1
-    if not out and key:
+    if len(out) == 0 and len(key) > 0:
         single = os.environ.get(key, '').strip()
-        if single:
+        if len(single) > 0:
             for part in single.replace(',', ' ').split():
                 out.append(part)
     return out
@@ -883,7 +867,7 @@ def _int_prefix(value: str) -> int | None:
             digits += char
         else:
             break
-    if not digits:
+    if len(digits) == 0:
         return None
     return int(digits)
 
@@ -902,401 +886,54 @@ def _normalize_time_unit(value: str) -> str:
     return 'hour'
 
 
-def ephemeris_params_from_env() -> EphemerisParams | None:
-    """Build EphemerisParams from CGI-style environment variables.
+from ephemeris_tools.params_env import (  # noqa: E402
+    ephemeris_params_from_env,
+    tracker_params_from_env,
+    viewer_params_from_env,
+)
 
-    Reads NPLANET, start, stop, interval, viewpoint, columns, etc. from env.
-
-    Returns:
-        EphemerisParams or None if required keys are missing/invalid.
-    """
-    nplanet_s = _get_env('NPLANET')
-    if not nplanet_s:
-        return None
-    try:
-        nplanet = int(nplanet_s.strip())
-    except ValueError as e:
-        logger.error('Invalid NPLANET %r (must be integer 4-9): %s', nplanet_s, e)
-        return None
-    if nplanet < 4 or nplanet > 9:
-        logger.error('NPLANET %d out of range (must be 4-9)', nplanet)
-        return None
-
-    start = _get_env('start') or _get_env('START_TIME')
-    stop = _get_env('stop') or _get_env('STOP_TIME')
-    if not start or not stop:
-        return None
-
-    interval_s = _get_env('interval', '1')
-    try:
-        interval = float(interval_s)
-    except ValueError as e:
-        logger.error('Invalid interval %r (must be number): %s; using 1.0', interval_s, e)
-        interval = 1.0
-    time_unit = _normalize_time_unit(_get_env('time_unit', 'hour'))
-
-    ephem_s = _get_env('ephem', '0')
-    try:
-        ephem_version = int(ephem_s.split()[0])
-    except (ValueError, IndexError) as e:
-        logger.error('Invalid ephem %r (must be integer): %s; using 0 (latest)', ephem_s, e)
-        ephem_version = 0
-
-    viewpoint = _get_env('viewpoint', 'observatory')
-    observatory = _get_env('observatory', "Earth's Center")
-    lat_s = _get_env('latitude')
-    lon_s = _get_env('longitude')
-    alt_s = _get_env('altitude')
-    lon_dir = _get_env('lon_dir', 'east')
-    try:
-        lat = float(lat_s) if lat_s else None
-    except ValueError:
-        logger.error('Invalid latitude %r: must be numeric', lat_s)
-        lat = None
-    try:
-        lon = float(lon_s) if lon_s else None
-    except ValueError:
-        logger.error('Invalid longitude %r: must be numeric', lon_s)
-        lon = None
-    try:
-        alt = float(alt_s) if alt_s else None
-    except ValueError:
-        logger.error('Invalid altitude %r: must be numeric', alt_s)
-        alt = None
-    if lon is not None and lon_dir.lower() == 'west':
-        lon = -lon
-
-    sc_traj_s = _get_env('sc_trajectory', '0')
-    try:
-        sc_trajectory = int(sc_traj_s[:4] or '0')
-    except ValueError as e:
-        logger.error('Invalid sc_trajectory %r: %s; using 0', sc_traj_s, e)
-        sc_trajectory = 0
-
-    column_strs = _get_keys_env('columns')
-    columns = parse_column_spec(column_strs) if column_strs else []
-
-    mooncol_strs = _get_keys_env('mooncols')
-    mooncols = parse_mooncol_spec(mooncol_strs) if mooncol_strs else []
-
-    moon_strs = _get_keys_env('moons')
-    from ephemeris_tools.planets import parse_moon_spec
-
-    moon_parsed = parse_moon_spec(nplanet, moon_strs) if moon_strs else []
-    moon_ids = [v if v >= 100 else 100 * nplanet + v for v in moon_parsed]
-
-    return EphemerisParams(
-        planet_num=nplanet,
-        start_time=start,
-        stop_time=stop,
-        interval=interval,
-        time_unit=time_unit,
-        ephem_version=ephem_version,
-        viewpoint=viewpoint,
-        observatory=observatory,
-        latitude_deg=lat,
-        longitude_deg=lon,
-        lon_dir=lon_dir,
-        altitude_m=alt,
-        sc_trajectory=sc_trajectory,
-        # In CGI mode we preserve exactly what the form submitted, so empty
-        # selections stay empty instead of injecting CLI defaults.
-        columns=columns,
-        mooncols=mooncols,
-        moon_ids=moon_ids,
-    )
-
-
-def viewer_params_from_env() -> ViewerParams | None:
-    """Build ``ViewerParams`` from CGI-style environment variables."""
-    nplanet_s = _get_env('NPLANET')
-    if not nplanet_s:
-        return None
-    try:
-        planet_num = int(nplanet_s)
-    except ValueError:
-        return None
-    time_str = _get_env('time')
-    if not time_str:
-        return None
-    fov_s = _get_env('fov', '1')
-    try:
-        fov_value = float(fov_s)
-    except ValueError:
-        # FORTRAN list-directed READ treats comma as a value separator, so
-        # strings like "557,000" parse as 557. Emulate that behavior.
-        head = fov_s.split(',', 1)[0].strip()
-        try:
-            fov_value = float(head)
-        except ValueError:
-            fov_value = 1.0
-    fov_unit = _get_env('fov_unit', 'degrees')
-
-    center_mode = _get_env('center', 'body')
-    if center_mode == 'J2000':
-        ra_type = _get_env('center_ra_type', 'hours').strip().lower()
-        is_ra_hours = not ra_type.startswith('d')
-        try:
-            ra_deg = _parse_sexagesimal_to_degrees(
-                _get_env('center_ra', '0'),
-                is_ra_hours=is_ra_hours,
-            )
-        except ValueError:
-            ra_deg = 0.0
-        try:
-            dec_deg = _parse_sexagesimal_to_degrees(
-                _get_env('center_dec', '0'),
-                is_ra_hours=False,
-            )
-        except ValueError:
-            dec_deg = 0.0
-        center = ViewerCenter(mode='J2000', ra_deg=ra_deg, dec_deg=dec_deg)
-    elif center_mode == 'ansa':
-        center = ViewerCenter(
-            mode='ansa',
-            ansa_name=_get_env('center_ansa') or None,
-            ansa_ew=_get_env('center_ew', 'east'),
-        )
-    elif center_mode == 'star':
-        center = ViewerCenter(mode='star', star_name=_get_env('center_star') or None)
-    else:
-        center = ViewerCenter(mode='body', body_name=_get_env('center_body') or None)
-
-    viewpoint = _get_env('viewpoint', 'observatory')
-    observer = Observer(name="Earth's Center")
-    viewpoint_display: str | None = None
-    if viewpoint == 'latlon':
-        lat_s = _get_env('latitude')
-        lon_s = _get_env('longitude')
-        alt_s = _get_env('altitude')
-        lon_dir = _get_env('lon_dir', 'east')
-        try:
-            lat = float(lat_s) if lat_s else None
-        except ValueError:
-            lat = None
-        try:
-            lon = float(lon_s) if lon_s else None
-        except ValueError:
-            lon = None
-        if lon is not None and lon_dir.lower() == 'west':
-            lon = -lon
-        try:
-            alt = float(alt_s) if alt_s else None
-        except ValueError:
-            alt = None
-        observer = Observer(latitude_deg=lat, longitude_deg=lon, lon_dir=lon_dir, altitude_m=alt)
-        if lat_s and lon_s and alt_s:
-            # FORTRAN captions preserve original CGI precision for lat/lon/alt text.
-            viewpoint_display = f'({lat_s}, {lon_s} {lon_dir}, {alt_s})'
-    elif viewpoint == 'observatory':
-        obs_name = _get_env('observatory', "Earth's Center")
-        coords = _parse_observatory_coords(obs_name)
-        if coords is None:
-            observer = Observer(name=obs_name)
-        else:
-            lat, lon, alt = coords
-            observer = Observer(
-                name=obs_name,
-                latitude_deg=lat,
-                longitude_deg=lon,
-                altitude_m=alt,
-            )
-    elif viewpoint:
-        observer = Observer(name=viewpoint)
-
-    moon_tokens = _get_keys_env('moons')
-    from ephemeris_tools.planets import parse_moon_spec
-
-    moon_ids = parse_moon_spec(planet_num, moon_tokens) if moon_tokens else None
-    rings_raw = _get_env('rings')
-    ring_names = None
-    if rings_raw:
-        ring_names = []
-        for comma_part in rings_raw.split(','):
-            for amp_part in comma_part.split('&'):
-                token = amp_part.strip()
-                if token:
-                    ring_names.append(token)
-    blank_flag = _get_env('blank', '').lower()
-    blank_disks = blank_flag in {'yes', 'y', 'true', '1'}
-    meridians_flag = _get_env('meridians', '').lower()
-    meridians = meridians_flag in {'yes', 'y', 'true', '1'}
-    opacity = _get_env('opacity', 'Transparent') or 'Transparent'
-    peris = _get_env('peris', 'None') or 'None'
-    peripts_s = _get_env('peripts', '4')
-    try:
-        peripts = float(peripts_s)
-    except ValueError:
-        peripts = 4.0
-    arcmodel = _get_env('arcmodel') or None
-    arcpts_s = _get_env('arcpts', '4')
-    try:
-        arcpts = float(arcpts_s)
-    except ValueError:
-        arcpts = 4.0
-    other_bodies = _get_keys_env('other')
-    labels = _get_env('labels', 'Small (6 points)')
-    moonpts_s = _get_env('moonpts', '0')
-    try:
-        moonpts = float(moonpts_s)
-    except ValueError:
-        moonpts = 0.0
-    title = _get_env('title')
-    additional_flag = _get_env('additional', '').lower()
-    show_standard_stars = additional_flag in {'yes', 'y', 'true', '1'}
-    extra_star: ExtraStar | None = None
-    if show_standard_stars:
-        extra_ra_s = _get_env('extra_ra', '')
-        extra_dec_s = _get_env('extra_dec', '')
-        extra_ra_type = _get_env('extra_ra_type', 'hours').strip().lower()
-        is_extra_ra_hours = not extra_ra_type.startswith('d')
-        if extra_ra_s.strip() and extra_dec_s.strip():
-            try:
-                extra_star = ExtraStar(
-                    name=_get_env('extra_name', ''),
-                    ra_deg=_parse_sexagesimal_to_degrees(
-                        extra_ra_s,
-                        is_ra_hours=is_extra_ra_hours,
-                    ),
-                    dec_deg=_parse_sexagesimal_to_degrees(
-                        extra_dec_s,
-                        is_ra_hours=False,
-                    ),
-                )
-            except ValueError:
-                extra_star = None
-    ephem_s = _get_env('ephem', '0')
-    ephem_value = ephem_s.split()[0] if ephem_s else '0'
-    try:
-        ephem_version = int(ephem_value)
-    except ValueError:
-        ephem_version = 0
-
-    display = ViewerDisplayInfo(
-        ephem_display=_get_env('ephem') or None,
-        moons_display=_get_env('moons') or None,
-        rings_display=_get_env('rings') or None,
-        viewpoint_display=viewpoint_display,
-    )
-    return ViewerParams(
-        planet_num=planet_num,
-        time_str=time_str,
-        fov_value=fov_value,
-        fov_unit=fov_unit,
-        center=center,
-        observer=observer,
-        ephem_version=ephem_version,
-        moon_ids=moon_ids,
-        ring_names=ring_names,
-        blank_disks=blank_disks,
-        opacity=opacity,
-        labels=labels,
-        moonpts=moonpts,
-        peris=peris,
-        peripts=peripts,
-        meridians=meridians,
-        arcmodel=arcmodel,
-        arcpts=arcpts,
-        torus=_get_env('torus', '').strip().lower() in {'yes', 'y', 'true', '1'},
-        torus_inc=_safe_float(_get_env('torus_inc', '6.8') or '6.8', 6.8),
-        torus_rad=_safe_float(_get_env('torus_rad', '422000') or '422000', 422000),
-        other_bodies=other_bodies if other_bodies else None,
-        show_standard_stars=show_standard_stars,
-        extra_star=extra_star,
-        title=title,
-        display=display,
-    )
-
-
-def tracker_params_from_env() -> TrackerParams | None:
-    """Build ``TrackerParams`` from CGI-style environment variables."""
-    nplanet_s = _get_env('NPLANET')
-    if not nplanet_s:
-        return None
-    try:
-        planet_num = int(nplanet_s)
-    except ValueError:
-        return None
-    start_time = _get_env('start')
-    stop_time = _get_env('stop')
-    if not start_time or not stop_time:
-        return None
-    interval_s = _get_env('interval', '1')
-    try:
-        interval = float(interval_s)
-    except ValueError:
-        interval = 1.0
-    time_unit = _normalize_time_unit(_get_env('time_unit', 'hour'))
-    viewpoint = _get_env('viewpoint', 'observatory')
-    observer = Observer(name="Earth's Center")
-    if viewpoint == 'observatory':
-        obs_name = _get_env('observatory', "Earth's Center")
-        coords = _parse_observatory_coords(obs_name)
-        if coords is None:
-            observer = Observer(name=obs_name)
-        else:
-            lat, lon, alt = coords
-            observer = Observer(
-                name=obs_name,
-                latitude_deg=lat,
-                longitude_deg=lon,
-                altitude_m=alt,
-            )
-    elif viewpoint == 'latlon':
-        lat_s = _get_env('latitude')
-        lon_s = _get_env('longitude')
-        alt_s = _get_env('altitude')
-        lat_deg: float | None
-        lon_deg: float | None
-        alt_m: float | None
-        try:
-            lat_deg = float(lat_s) if lat_s else None
-        except ValueError:
-            lat_deg = None
-        try:
-            lon_deg = float(lon_s) if lon_s else None
-        except ValueError:
-            lon_deg = None
-        try:
-            alt_m = float(alt_s) if alt_s else None
-        except ValueError:
-            alt_m = None
-        if lon_deg is not None and _get_env('lon_dir', 'east').lower() == 'west':
-            lon_deg = -lon_deg
-        observer = Observer(latitude_deg=lat_deg, longitude_deg=lon_deg, altitude_m=alt_m)
-    elif viewpoint:
-        observer = Observer(name=viewpoint)
-    moon_tokens = _get_keys_env('moons')
-    from ephemeris_tools.planets import parse_moon_spec
-
-    moon_ids = parse_moon_spec(planet_num, moon_tokens) if moon_tokens else []
-    rings_raw = _get_keys_env('rings')
-    ring_names = [r.strip() for r in rings_raw if r.strip()] if rings_raw else None
-    xrange_s = _get_env('xrange')
-    try:
-        xrange = float(xrange_s) if xrange_s else None
-    except ValueError:
-        xrange = None
-    xunit_raw = _get_env('xunit', 'arcsec')
-    xunit = 'radii' if 'radii' in xunit_raw.lower() else 'arcsec'
-    title = _get_env('title')
-    ephem_s = _get_env('ephem', '0')
-    ephem_value = ephem_s.split()[0] if ephem_s else '0'
-    try:
-        ephem_version = int(ephem_value)
-    except ValueError:
-        ephem_version = 0
-    return TrackerParams(
-        planet_num=planet_num,
-        start_time=start_time,
-        stop_time=stop_time,
-        interval=interval,
-        time_unit=time_unit,
-        observer=observer,
-        ephem_version=ephem_version,
-        moon_ids=moon_ids,
-        ring_names=ring_names,
-        xrange=xrange,
-        xunit=xunit,
-        title=title,
-    )
+__all__ = [
+    'COL_LSEP',
+    'COL_MJD',
+    'COL_OBSDIST',
+    'COL_PHASE',
+    'COL_RADDEG',
+    'COL_RADEC',
+    'COL_RADIUS',
+    'COL_SUBOBS',
+    'COL_SUBSOL',
+    'COL_SUNOPEN',
+    'COL_SUNRD',
+    'COL_SUNSEP',
+    'COL_YDHM',
+    'COL_YDHMS',
+    'COL_YMDHM',
+    'COL_YMDHMS',
+    'MCOL_OFFDEG',
+    'MCOL_OFFSET',
+    'MCOL_ORBLON',
+    'MCOL_ORBOPEN',
+    'MCOL_PHASE',
+    'MCOL_RADEC',
+    'MCOL_SUBOBS',
+    'MCOL_SUBSOL',
+    'PLANET_NAME_TO_NUM',
+    'EphemerisParams',
+    'ExtraStar',
+    'Observer',
+    'TrackerParams',
+    'ViewerCenter',
+    'ViewerDisplayInfo',
+    'ViewerParams',
+    'ephemeris_params_from_env',
+    'parse_center',
+    'parse_column_spec',
+    'parse_fov',
+    'parse_mooncol_spec',
+    'parse_observer',
+    'parse_planet',
+    'parse_ring_spec',
+    'parse_viewer_rings',
+    'tracker_params_from_env',
+    'viewer_params_from_env',
+]
