+++ stubs/cspyce/__init__.pyi
@@ -17,6 +17,7 @@ def bodvar(bodyid: int, item: str) -> tuple[float, ...]: ...
 def bodfnd(bodyid: int, item: str) -> bool: ...
 def bodvrd(body: str, item: str) -> tuple[float, ...]: ...
 def bodc2n(bodyid: int) -> str: ...
+def bodn2c(name: str) -> int: ...
 def tipbod(ref: str, body: int, et: float) -> tuple[tuple[float, float, float], ...]: ...
 def spkez(
     targ: int,
@@ -32,6 +33,13 @@ def twovec(
     ax2: int,
 ) -> tuple[tuple[float, float, float], ...]: ...
 def recrad(rectan: tuple[float, float, float] | Any) -> tuple[float, float, float]: ...
+def radrec(
+    r: float, colat: float, lon: float
+) -> tuple[float, float, float]: ...
+def vperp(
+    v1: tuple[float, float, float] | Any,
+    v2: tuple[float, float, float] | Any,
+) -> tuple[float, float, float]: ...
 def vminus(v: tuple[float, float, float] | Any) -> tuple[float, float, float]: ...
 def vsep(v1: tuple[float, float, float] | Any, v2: tuple[float, float, float] | Any) -> float: ...
 def vnorm(v: tuple[float, float, float] | Any) -> float: ...
