+++ stubs/julian/__init__.pyi
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+def set_ut_model(model: str) -> None: ...
 def load_lsk(path: str | None = None) -> None: ...
 def day_sec_from_string(string: str) -> tuple[int, float]: ...
 def tai_from_day_sec(day: int, sec: float) -> float: ...
